refactor: reservationtime을 별도의 패캐지로 관리하지 않도록 변경
