+++ src/main/java/roomescape/user/reservation/domain/ReservationService.java
@@ -6,8 +6,6 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.user.reservationtime.domain.ReservationTime;
-import roomescape.user.reservationtime.domain.ReservationTimeRepository;
 
 @RequiredArgsConstructor
 @Service
