+++ src/main/java/roomescape/user/reservation/presentation/UserReservationTimeRestController.java
@@ -1,4 +1,4 @@
-package roomescape.user.reservationtime.controller;
+package roomescape.user.reservation.presentation;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.user.reservationtime.domain.ReservationTime;
-import roomescape.user.reservationtime.domain.ReservationTimeRepository;
+import roomescape.user.reservation.domain.ReservationTime;
+import roomescape.user.reservation.domain.ReservationTimeRepository;
 
 @RequiredArgsConstructor
 @RestController
