+++ src/test/java/roomescape/user/reservationtime/domain/FakeReservationTimeRepositoryTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import roomescape.fake.FakeReservationTimeRepository;
+import roomescape.user.reservation.domain.ReservationTime;
 
 class FakeReservationTimeRepositoryTest {
